SectionManager: fill_sections -> SectionFiller
